// seductx/src/app/model/escort-agent.model.ts
export interface EscortAgent {
  escort_id: string;
  first_name: string;
  last_name: string;
  profile_image: string;
  image1: string;
  image2: string;
  image3: string;
  image4: string;
  image5: string;
  image6: string;
  reset_code: string;
  country_code: string;
  phone_number: string;
  age: string;
  height: string;
  status: string;
  progress_status: string;
  country_id: string;
  country_name: string;
  body_type_id: string;
  body_type_name: string;
  city_id: string;
  city_name: string;
  gender_id: string;
  gender: string;
  bust_size_id: string;
  bust_size: string;
  cup_size_id: string;
  cup_size: string;
  waise_size_id: string;
  waise_size: string;
  hip_size_id: string;
  hip_size: string;
  english_id: string;
  english_type: string;
  chinese_id: string;
  chinese_type: string;
  japanese_id: string;
  japanese_type: string;
  korean_id: string;
  korean_type: string;
  distance: string;
  online_status: string;
  avg_rating: string;
}